mem: Implemented reading and writing
For now we'll just let it segfault on bad addresses. Later, we'll actually
catch these bad addresses cleanly.

+++ src/mem.h
@@ -13,6 +13,6 @@ uint8_t  mem_read8(uint32_t loc);
 uint16_t mem_read16(uint32_t loc);
 uint32_t mem_read32(uint32_t loc);
 
-int mem_write8(uint32_t loc, uint8_t data);
-int mem_write16(uint32_t loc, uint16_t data);
-int mem_write32(uint32_t loc, uint32_t data);
+uint8_t mem_write8(uint32_t loc, uint8_t data);
+uint16_t mem_write16(uint32_t loc, uint16_t data);
+uint32_t mem_write32(uint32_t loc, uint32_t data);
